# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CyberAI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ThreatSeverity = typeof ThreatSeverity[keyof typeof ThreatSeverity];


export const ThreatSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface Threat {
  id: number;
  title: string;
  description: string;
  severity: ThreatSeverity;
  category: string;
  affectedSystems?: string[];
  /** @nullable */
  cveId?: string | null;
  publishedAt: string;
  source: string;
  /** @nullable */
  sourceUrl?: string | null;
  mitigations?: string[];
}

export type ThreatsSummaryBySeverity = {[key: string]: number};

export type ThreatsSummaryByCategory = {[key: string]: number};

export interface ThreatsSummary {
  total: number;
  recentCount: number;
  bySeverity: ThreatsSummaryBySeverity;
  byCategory: ThreatsSummaryByCategory;
}

export type RecommendationCategory = typeof RecommendationCategory[keyof typeof RecommendationCategory];


export const RecommendationCategory = {
  antivirus: 'antivirus',
  firewall: 'firewall',
  vpn: 'vpn',
  'password-manager': 'password-manager',
  'endpoint-protection': 'endpoint-protection',
  'threat-intelligence': 'threat-intelligence',
  siem: 'siem',
  'vulnerability-scanner': 'vulnerability-scanner',
} as const;

export interface Recommendation {
  id: number;
  name: string;
  description: string;
  category: RecommendationCategory;
  platforms: string[];
  rating: number;
  price: string;
  isFree?: boolean;
  isFeatured: boolean;
  /** @nullable */
  websiteUrl?: string | null;
  pros?: string[];
  cons?: string[];
  bestFor?: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessageInput {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}

export type ListThreatsParams = {
category?: string;
severity?: string;
};

export type ListRecommendationsParams = {
category?: string;
platform?: string;
};