# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CyberAI API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: threats
    description: Cybersecurity threat research
  - name: recommendations
    description: Antivirus and security program recommendations
  - name: openai
    description: OpenAI AI chat operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /threats:
    get:
      operationId: listThreats
      tags: [threats]
      summary: List recent cybersecurity threats
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: severity
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of threats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Threat"

  /threats/summary:
    get:
      operationId: getThreatsSummary
      tags: [threats]
      summary: Get threat statistics summary
      responses:
        "200":
          description: Threat summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreatsSummary"

  /threats/{id}:
    get:
      operationId: getThreat
      tags: [threats]
      summary: Get a single threat by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Threat detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Threat"
        "404":
          description: Not found

  /recommendations:
    get:
      operationId: listRecommendations
      tags: [recommendations]
      summary: List security program recommendations
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: platform
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recommendation"

  /recommendations/featured:
    get:
      operationId: getFeaturedRecommendations
      tags: [recommendations]
      summary: Get top featured security programs
      responses:
        "200":
          description: Featured recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recommendation"

  /recommendations/{id}:
    get:
      operationId: getRecommendation
      tags: [recommendations]
      summary: Get a single recommendation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recommendation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendation"
        "404":
          description: Not found

  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"

  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"

  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenaiMessageInput"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Threat:
      type: object
      required: [id, title, description, severity, category, publishedAt, source]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, low]
        category:
          type: string
        affectedSystems:
          type: array
          items:
            type: string
        cveId:
          type: ["string", "null"]
        publishedAt:
          type: string
        source:
          type: string
        sourceUrl:
          type: ["string", "null"]
        mitigations:
          type: array
          items:
            type: string

    ThreatsSummary:
      type: object
      required: [total, bySeverity, byCategory, recentCount]
      properties:
        total:
          type: integer
        recentCount:
          type: integer
        bySeverity:
          type: object
          additionalProperties:
            type: integer
        byCategory:
          type: object
          additionalProperties:
            type: integer

    Recommendation:
      type: object
      required: [id, name, description, category, platforms, rating, price, isFeatured]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [antivirus, firewall, vpn, password-manager, endpoint-protection, threat-intelligence, siem, vulnerability-scanner]
        platforms:
          type: array
          items:
            type: string
        rating:
          type: number
        price:
          type: string
        isFree:
          type: boolean
        isFeatured:
          type: boolean
        websiteUrl:
          type: ["string", "null"]
        pros:
          type: array
          items:
            type: string
        cons:
          type: array
          items:
            type: string
        bestFor:
          type: string

    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt

    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt

    OpenaiConversationInput:
      type: object
      properties:
        title:
          type: string
      required:
        - title

    OpenaiMessageInput:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages

    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error