# lib/api-zod/src/generated/types/openaiMessage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CyberAI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: Date;
}